+++ src/components/PeopleTable/PeopleTable.tsx
@@ -1,11 +1,15 @@
+import { useParams } from 'react-router-dom';
 import { Person } from '../../types';
 import { PeopleLink } from '../PeopleLink';
+import classNames from 'classnames';
 
 type Props = {
   people: Person[];
 };
 
 export const PeopleTable: React.FC<Props> = ({ people }) => {
+  const { slug } = useParams();
+
   return (
     <table
       data-cy="peopleTable"
@@ -23,9 +27,46 @@ export const PeopleTable: React.FC<Props> = ({ people }) => {
       </thead>
 
       <tbody>
-        {people.map(person => (
-          <PeopleLink person={person} people={people} key={person.name} />
-        ))}
+        {people.map(person => {
+          const mother = people.find(p => p.name === person.motherName);
+          const father = people.find(p => p.name === person.fatherName);
+
+          return (
+            <tr
+              data-cy="person"
+              className={classNames({
+                'has-background-warning': slug === person.slug,
+              })}
+              key={person.name}
+            >
+              <td>
+                <PeopleLink person={person} />
+              </td>
+
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+              <td>
+                {!person.motherName ? (
+                  '-'
+                ) : mother ? (
+                  <PeopleLink person={mother} />
+                ) : (
+                  person.motherName
+                )}
+              </td>
+              <td>
+                {!person.fatherName ? (
+                  '-'
+                ) : father ? (
+                  <PeopleLink person={father} />
+                ) : (
+                  person.fatherName
+                )}
+              </td>
+            </tr>
+          );
+        })}
       </tbody>
     </table>
   );
